+++ .jupyter/.ipython/profile_default/startup/start.py
@@ -1,14 +1,17 @@
 import os
 import sys
+
 sys.path.append('/home/contuser1/project/module')
 
 import logging
-import numpy as np
-import pandas as pd
-from tqdm import tqdm
+
 import const
 import matplotlib.pyplot as plt
-from IPython.display import display, HTML
+import numpy as np
+import pandas as pd
+from IPython.display import HTML, display
+from tqdm import tqdm
+
 plt.style.use(['fivethirtyeight'])
 
 # not use gpu
