+++ Pipfile
@@ -22,6 +22,7 @@ cssselect = "*"
 tqdm = "*"
 pydantic = "*"
 autopep8 = "*"
+isort = "*"
 
 [requires]
 python_version = "3.7"
