+++ src/domain/const.py
@@ -1,6 +1,7 @@
 # python 3.7.5
 # coding: utf-8
 import os
+
 # TODO お前はインフラだ
 
 if os.getenv('MYSQL_ENV') == "TEST":
