+++ src/main.py
@@ -4,27 +4,15 @@
 MYSQLへ公式データを格納する
 """
 import argparse
-from logging import (
-    getLogger,
-    Formatter,
-    INFO,
-    StreamHandler,
-)
+from logging import INFO, Formatter, StreamHandler, getLogger
 
 import coloredlogs
 
-from application.usecase import BoatRaceUsecase
 from application.argument import Options
-from domain.const import (
-    CL_FIELD_STYLES,
-    CL_LEVEL_STYLES,
-    DATE_FMT,
-    FMT,
-    MODULE_LOG_NAME,
-    MAIN_LOGNAME
-)
+from application.usecase import BoatRaceUsecase
+from domain.const import (CL_FIELD_STYLES, CL_LEVEL_STYLES, DATE_FMT, FMT,
+                          MAIN_LOGNAME, MODULE_LOG_NAME)
 from module.log import TqdmLoggingHandler
-
 
 if __name__ == '__main__':
     # logging設定
