+++ src/module/getdata_lxml.py
@@ -3,13 +3,13 @@
 """
 HTMLから情報をスクレイピングするためのモジュール
 """
+import re
+import sys
+import time
 from dataclasses import asdict
 from datetime import datetime, timedelta
 from logging import getLogger
 from pathlib import Path
-import re
-import sys
-import time
 from typing import Iterator
 from urllib.request import urlopen
 
@@ -18,11 +18,8 @@ import numpy as np
 import pandas as pd
 
 from domain import const
-from domain.model.info import (
-    Tansho, ThreeRenfuku,
-    ThreeRentan, TwoRenfuku,
-    TwoRentan
-)
+from domain.model.info import (Tansho, ThreeRenfuku, ThreeRentan, TwoRenfuku,
+                               TwoRentan)
 from module.getter import GetParserContent
