+++ src/module/log.py
@@ -1,4 +1,5 @@
 import logging
+
 import tqdm
