+++ src/module/test/testdt2mysql/test_jd2sql.py
@@ -6,6 +6,7 @@ jyodata2sqlテスト
 import pytest
 
 from module.dt2sql import JyoData2sql
+
 from ..common import CommonMethod
