+++ src/module/test/testdt2mysql/test_master2sql.py
@@ -4,8 +4,10 @@
 master2sqlモジュール用単体テスト
 """
 import pytest
-from ..common import CommonMethod
+
 from module.master2sql import JyoMaster2sql
+
+from ..common import CommonMethod
 
 
 @pytest.mark.run(order=1)
