+++ src/module/test/testdt2mysql/test_odds2sql.py
@@ -5,16 +5,11 @@ odds2sqlテスト
 """
 import pytest
 
-
-from domain.model.info import (
-    Tansho,
-    ThreeRenfuku,
-    ThreeRentan,
-    TwoRenfuku,
-    TwoRentan
-)
+from domain.model.info import (Tansho, ThreeRenfuku, ThreeRentan, TwoRenfuku,
+                               TwoRentan)
 from module.dt2sql import Odds2sql
 from module.getdata_lxml import OfficialOdds
+
 from ..common import CommonMethod
