+++ src/module/test/testgetdata/test_officialchokuzen.py
@@ -4,8 +4,10 @@
 getdataモジュール用単体テスト
 """
 import pytest
+
 from module.getdata import OfficialChokuzen
 from module.getter import GetParserContent
+
 from ..common import CommonMethod
