+++ src/module/test/testgetdata/test_officialprogram.py
@@ -5,8 +5,10 @@ getdataモジュール用単体テスト
 """
 
 import pytest
+
 from module.getdata import OfficialProgram
 from module.getter import GetParserContent
+
 from ..common import CommonMethod
