+++ src/module/test/testgetdata_lxml/test_getholdplace.py
@@ -4,8 +4,10 @@
 getdataモジュール用単体テスト
 """
 import pytest
+
 from module.getdata_lxml import GetHoldPlacePast
 from module.getter import GetParserContent
+
 from ..common import CommonMethod
